🎨 Move search logic to model & util
