+++ .sequelizerc
@@ -1,3 +1,5 @@
+require('babel-register');
+
 const path = require('path');
 
 module.exports = {
