+++ server/models/asset.js
@@ -1,4 +1,10 @@
+import { RESULT_PER_PAGE } from '../../constant';
+import { ValidationError } from './validator';
+import { parseKeyword } from '../util/paginator';
+
 module.exports = (sequelize, DataTypes) => {
+  const { Op } = sequelize.Sequelize;
+
   const asset = sequelize.define(
     'asset',
     {
@@ -51,5 +57,65 @@ module.exports = (sequelize, DataTypes) => {
       timestamps: true,
     }
   );
+
+  asset.searchByKeyword = async (models, q, offset) => {
+    const [keywords, tags, users] = parseKeyword(q);
+    const tagWhere = {};
+    const userWhere = {};
+    const where = {};
+    let tagRequired = false;
+    let userRequired = false;
+    if (!keywords.length && !tags.length && !users.length) {
+      throw new ValidationError('no search keyword');
+    }
+    if (keywords.length) {
+      where.description = { [Op.iLike]: { [Op.any]: keywords } };
+    }
+    if (tags.length) {
+      tagRequired = true;
+      tagWhere.name = { [Op.iLike]: { [Op.any]: tags } };
+    }
+    if (users.length) {
+      userRequired = true;
+      userWhere[Op.or] = {
+        displayName: { [Op.iLike]: { [Op.any]: users } },
+        likecoinId: { [Op.iLike]: { [Op.any]: users } },
+      };
+    }
+
+    return models.asset.findAndCountAll({
+      distinct: true,
+      include: [
+        {
+          as: 'like',
+          attributes: [['total_like', 'count']],
+          model: models.assetLike,
+          order: [['totalLike', 'DESC']],
+          // required: true, // This should be enabled but caused bug when sorting
+        },
+        {
+          attributes: ['name'],
+          model: models.tag,
+          required: tagRequired,
+          through: { attributes: [] },
+          where: tagWhere,
+        },
+        {
+          attributes: ['displayName', 'likecoinId'],
+          model: models.user,
+          required: userRequired,
+          where: userWhere,
+        },
+      ],
+      limit: RESULT_PER_PAGE,
+      offset,
+      order: [
+        [{ as: 'like', model: models.assetLike }, 'totalLike', 'DESC'],
+        ['updatedAt', 'DESC'],
+      ],
+      where,
+    });
+  };
+
   return asset;
 };
